UI更新, 敵グラフロード変更
敵グラフのファイルが存在しない場合変更するようにした

+++ Assets/Code/EnemyGraphLoader.cs
@@ -8,7 +8,7 @@ public class EnemyGraphLoader : MonoBehaviour
     public const string graphExtension = ".json";
     public const string graphPath = graphDirectory + "\\" + graphFile + graphExtension;
 
-    
+
     void Start()
     {
         if (!Directory.Exists(Application.persistentDataPath.Replace("/", "\\") + $"\\{graphDirectory}"))
@@ -16,11 +16,14 @@ public class EnemyGraphLoader : MonoBehaviour
             Directory.CreateDirectory(Application.persistentDataPath.Replace("/", "\\") + $"\\{graphDirectory}");
         }
 
-        TextAsset json = Resources.Load<TextAsset>(graphDirectory+"/"+graphFile);
-
-        string data = json?.text;
         string path = Application.persistentDataPath.Replace("/", "\\") + $"\\{graphPath}";
+        if (!File.Exists(path))
+        {
+            TextAsset json = Resources.Load<TextAsset>(graphDirectory + "/" + graphFile);
 
-        File.WriteAllText(path, data, System.Text.Encoding.UTF8);
+            string data = json?.text;
+
+            File.WriteAllText(path, data, System.Text.Encoding.UTF8);
+        }
     }
 }
